refactor: extract useEffect logic into useEmpresaData hook
- Removed internal data loading effect from Table
- Hook now handles fetching empresa and role info cleanly

See #16
